refactor: extract velociraptor routes into routes/velociraptor.ts
Pure structural move of the 27 Velociraptor endpoints (/velociraptor/* and
/cases/:id/velociraptor/*) out of createApp; no handler logic changed.

The stateful machinery driving these routes STAYS in createApp because its
resume functions (resumeVeloMonitors / resumeVeloHuntStatusPolls) run once at
startup and from POST /velociraptor/reconnect — a route module can't be invoked
at boot. The self-rescheduling monitor/hunt-status timer loops, their timer Maps
(veloMonitorTimers / veloStatusTimers, kept fully private), and the hunt-collect
/ ingest cores therefore remain in createApp; the moved routes reach them through
graduated RouteContext members: refreshVeloClients, resumeVelo*, schedule/poll/
stop VeloMonitor, schedule/poll VeloHuntStatus, importVeloHuntResults,
ingestVeloArtifactMap / ingestVeloUploads, createVeloMonitor, recordHuntDeploy
(stable methods) plus the veloHuntTimers() live accessor (run-bundle/deploy-hunt
set a fixed-delay collect timer on that shared Map). rebuildVelo + collectHostResolved
(route-only helpers) moved into the module. Dead imports trimmed.
